requirements.txt
matplotlib
pandas
torch
numpy
scikit-learn
Pillow
albumentations
segmentation-models-pytorch
